Conflicts:
	app/cli/corrupto.php
	app/functions.php

+++ app/cli/sitemap.php
@@ -34,8 +34,12 @@ function images()
     foreach($db->getCollection('corruptos')->find() as $corrupto) {
         $ext = explode('.', $corrupto->image);
         $ext = strtolower(end($ext));
-        `wget "{$corrupto->image}" -O "{$base}/{$corrupto->id}.{$ext}"`;
-        `wget "{$corrupto->avatar}" -O "{$base}/{$corrupto->id}:large.{$ext}"`;
+        if (!is_file("{$base}/{$corrupto->id}.{$ext}")) {
+            `wget "{$corrupto->image}" -O "{$base}/{$corrupto->id}.{$ext}"`;
+        }
+        if (!is_file("{$base}/{$corrupto->id}:large.{$ext}")) {
+            `wget "{$corrupto->avatar}" -O "{$base}/{$corrupto->id}:large.{$ext}"`;
+        }
     }
     `cd $base ; jpegoptim --strip-all * ; chmod +r *`;
 }
